Fixed warnings. JIRA: CXX-4752
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@63641 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/data_loaders/genbank/info_cache.cpp
@@ -406,7 +406,7 @@ void CInfoCache_Base::SetMaxGCQueueSize(size_t max_size)
 {
     TCacheMutex::TWriteLockGuard guard(m_CacheMutex);
     m_MaxGCQueueSize = max_size;
-    m_MinGCQueueSize = m_MaxGCQueueSize*0.9;
+    m_MinGCQueueSize = size_t(m_MaxGCQueueSize*0.9);
     if ( m_CurGCQueueSize > m_MaxGCQueueSize ) {
         x_GC();
     }
